Fri, Dec 9, 2022, 9:52 PM +00:00

+++ pset1/mario/calculator.c
@@ -3,8 +3,8 @@
 
 int main(void)
 {
-    int x = get_int("x: ");
-    int y = get_int("y: ");
-    int z = x + y
-    printf("%i\n", z);
+    long x = get_int("x: ");
+    long y = get_int("y: ");
+    long z = x + y
+    printf("%li\n", z);
 }
